fmcg-saas-app/requirements.txt
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
openpyxl==3.1.2
xlrd==2.0.1
prophet==1.1.5
xgboost==2.0.3
scikit-learn==1.4.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
groq==0.4.2
plotly==5.18.0
python-dotenv==1.0.0
supabase==2.3.4
